@@ -1,50 +1,110 @@
 {
-  "name": "Migration-React",
-  "version": "1.0.0",
-  "description": "",
-  "main": "webpack.config.js",
-  "scripts": {
-    "build": "NODE_ENV=production",
-    "clean": "rimraf dist/",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "axios": "^0.12.0",
-    "babel-core": "^6.9.0",
-    "babel-loader": "^6.2.4",
-    "babel-preset-react": "^6.5.0",
-    "coffee-loader": "^0.7.2",
-    "css-loader": "^0.23.1",
-    "extract-text-webpack-plugin": "^0.9.0",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.17.0",
-    "loader-utils": "^0.2.15",
-    "material-ui": "^0.15.0",
-    "node-sass": "^3.7.0",
-    "path": "^0.12.7",
-    "sass-loader": "^3.2.0",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.16.1"
-  },
+  "name": "updatedmigrationtrack-ui",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "babel-preset-es2015": "^6.9.0",
-    "classnames": "^2.2.5",
-    "leaflet": "^0.7.7",
-    "leaflet.markercluster": "^0.5.0",
-    "loader-utils": "^0.2.15",
-    "moment": "^2.13.0",
-    "moment-duration-format": "^1.3.0",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "react-leaflet": "^0.12.2",
+    "autoprefixer": "7.1.6",
+    "axios": "^0.17.1",
+    "babel-core": "6.26.0",
+    "babel-eslint": "7.2.3",
+    "babel-jest": "20.0.3",
+    "babel-loader": "7.1.2",
+    "babel-preset-react-app": "^3.1.0",
+    "babel-runtime": "6.26.0",
+    "case-sensitive-paths-webpack-plugin": "2.1.1",
+    "chalk": "1.1.3",
+    "css-loader": "0.28.7",
+    "dotenv": "4.0.0",
+    "eslint": "4.10.0",
+    "eslint-config-react-app": "^2.0.1",
+    "eslint-loader": "1.9.0",
+    "eslint-plugin-flowtype": "2.39.1",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.4.0",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.5",
+    "fs-extra": "3.0.1",
+    "html-webpack-plugin": "2.29.0",
+    "jest": "20.0.4",
+    "leaflet": "^1.2.0",
+    "leaflet.markercluster": "^1.2.0",
+    "material-ui": "^0.20.0",
+    "moment": "^2.20.1",
+    "moment-duration-format": "^2.1.1",
+    "object-assign": "4.1.1",
+    "promise": "8.0.1",
+    "raf": "3.4.0",
+    "react": "^16.2.0",
+    "react-dev-utils": "^4.2.1",
+    "react-dom": "^16.2.0",
+    "react-leaflet": "^1.7.8",
     "react-leaflet-cluster-layer": "0.0.3",
-    "react-leaflet-control": "^1.1.2",
+    "react-leaflet-control": "^1.4.1",
     "react-leaflet-geojson-cluster": "^0.1.8",
-    "react-router": "^2.4.1",
-    "react-tap-event-plugin": "^1.0.0"
+    "react-tap-event-plugin": "^3.0.2",
+    "style-loader": "0.19.0",
+    "sw-precache-webpack-plugin": "0.11.4",
+    "url-loader": "0.6.2",
+    "webpack": "3.8.1",
+    "webpack-dev-server": "2.9.4",
+    "webpack-manifest-plugin": "1.3.2",
+    "whatwg-fetch": "2.0.3"
+  },
+  "scripts": {
+    "start": "node scripts/start.js",
+    "build": "node scripts/build.js",
+    "test": "node scripts/test.js --env=jsdom"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx,mjs}"
+    ],
+    "setupFiles": [
+      "<rootDir>/config/polyfills.js"
+    ],
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs}",
+      "<rootDir>/src/**/?(*.)(spec|test).{js,jsx,mjs}"
+    ],
+    "testEnvironment": "node",
+    "testURL": "http://localhost",
+    "transform": {
+      "^.+\\.(js|jsx|mjs)$": "<rootDir>/node_modules/babel-jest",
+      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
+      "^(?!.*\\.(js|jsx|mjs|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
+    },
+    "transformIgnorePatterns": [
+      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"
+    ],
+    "moduleNameMapper": {
+      "^react-native$": "react-native-web"
+    },
+    "moduleFileExtensions": [
+      "web.js",
+      "mjs",
+      "js",
+      "json",
+      "web.jsx",
+      "jsx",
+      "node"
+    ]
+  },
+  "babel": {
+    "presets": [
+      "react-app"
+    ]
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "devDependencies": {
+    "postcss-cssnext": "^3.0.2",
+    "postcss-each": "^0.10.0",
+    "postcss-flexbugs-fixes": "3.2.0",
+    "postcss-import": "^11.0.0",
+    "postcss-loader": "^2.0.8",
+    "postcss-mixins": "^6.2.0",
+    "react-router": "3.2.0"
   }
 }
