@@ -1,18 +1,18 @@
 import axios from "axios"
 import env from "./env"
 
-module.exports = {
+const api = {
   birds: function() {
     return axios({
     method: 'get',
-    url: env.URL_ROOT + '/api/birds',
+    url: env.URL_ROOT + '/birds',
     headers: {'Content-Type': 'application/json'}
     })
   },
   checkUserLoggedIn: function() {
     return axios({
     method: 'head',
-    url: env.URL_ROOT + '/api/user',
+    url: env.URL_ROOT + '/user',
     withCredentials: true,
     headers: {'Content-Type': 'application/json'}
     })
@@ -20,7 +20,7 @@ module.exports = {
   getUser: function() {
     return axios({
     method: 'get',
-    url: env.URL_ROOT + '/api/user',
+    url: env.URL_ROOT + '/user',
     withCredentials: true,
     headers: {'Content-Type': 'application/json'}
     })
@@ -28,14 +28,14 @@ module.exports = {
   getLocation: function(city) {
     return axios({
     method: 'get',
-    url: env.URL_ROOT + '/api/location/' + city,
+    url: env.URL_ROOT + '/location/' + city,
     headers: {'Content-Type': 'application/json'}
     })
   },
   saveBird: function(data) {
     return axios({
       method: 'post',
-      url: env.URL_ROOT + '/api/birds',
+      url: env.URL_ROOT + '/birds',
       data: data,
       withCredentials: true,
       headers: {'Content-Type': 'application/json'}
@@ -44,8 +44,10 @@ module.exports = {
   signOut: function() {
     return axios({
     method: 'get',
-    url: env.URL_ROOT + '/api/logout',
+    url: env.URL_ROOT + '/logout',
     headers: {'Content-Type': 'application/json'}
     })
   }
 }
+
+export default api;
